Move project-member under project
