+++ src/components/project/index.ts
@@ -2,3 +2,4 @@ export * from './dto';
 export * from './project.resolver';
 export * from './project.service';
 export * from './project.module';
+export * from './project-member';
