+++ src/components/project/project-member/dto/index.ts
@@ -2,4 +2,4 @@ export * from './role.dto';
 export * from './project-member.dto';
 export * from './create-project-member.dto';
 export * from './update-project-member.dto';
-export * from './list-projectMember.dto';
+export * from './list-project-members.dto';
