+++ src/components/project/project-member/dto/project-member.dto.ts
@@ -1,7 +1,7 @@
 import { DateTime } from 'luxon';
 import { Field, ObjectType } from 'type-graphql';
-import { DateTimeField, Resource } from '../../../common';
-import { SecuredUser } from '../../user';
+import { DateTimeField, Resource } from '../../../../common';
+import { SecuredUser } from '../../../user';
 import { SecuredRoles } from './role.dto';
 
 @ObjectType({
