+++ src/components/project/project.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { ProjectMemberModule } from '../project-member';
 import { InternshipProjectResolver } from './internship-project.resolver';
+import { ProjectMemberModule } from './project-member';
 import { ProjectResolver } from './project.resolver';
 import { ProjectService } from './project.service';
 import { TranslationProjectResolver } from './translation-project.resolver';
